First commit for the script
In the current state, this script is able to identify the two first rows of the labyrinth. Wrizz boss is not implemented yet.
Requirements are a rooted phone still, but I think it's possible to change that, I just need to do some more testing.
If you have ADB on your Machine installed, as well as a rooted phone that has a 1920X1080 resolution, feel free to give this script a try!